agent/requirements.txt
strands-agents==0.1.0
requests==2.32.4
python-dateutil==2.9.0
pytz==2024.1
pytest==8.3.3
pytest-mock==3.14.0